Add command to download inlets(pro)
Added a command to download inlets, and a flag for making
that download intlets-pro. We work out what arch we are
working on, and the os. We then get latest version, or
use the version supplied in --download-version to get the
correct binary from github.

This has been tested by running the commands on Linux arm64
and armhf (Rpi3)

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>