Fix issue in inlets-pro userdata
The name changed from having a suffix for linux, to having no
suffix at all. Also the version is bumped to 0.4.3. The issue
was found whilst trying to setup inlets-pro with inletsctl. The
operator may also need a patch for this change.

Tested by running the command as root on a Linux PC, before it
failed, after it worked.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>