Various fixes for ec2 provider create command
The create command failed when a default VPC was not
present for the ec2 provider. When the failure occurred there
was also a side effect of dangling security groups that
got created but not cleaned up in the event of failure.

Two additional arguments were added, vpc-id and subnet-id,
to mitigate the failure while also giving users the option
to choose a specific VPC and subnet if more than one VPC
exists. New validation checks that if one of these new arguments
are specified then the other must be as well.

Additional logic was added so that in the event of ec2
instance creation failure any security groups created
will get deleted.

Signed-off-by: Shane Witbeck <shane@digitalsanctum.com>