Move to secure defaults
Whilst inlets OSS can be useful for understanding how
tunnels work, it does not encrypt the control-plane by
default, so is not secure by default. This moves to using
inlets PRO to encrypt traffic over the tunnel.

Significant contributors to inletsctl are welcome to a year's
free license for non-commercial use for inlets PRO.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>