Accommodate Session Tokens for ec2
Adds the capability for users to use temporary credentials
 when using ec2.  To do this there needed to be provision for
the session token to be passed.  This has been added both as
a CLI parameter and a file (following the same pattern as the
Access key and secret key).

When calling `getFileOrString` for  sessionToken the required
flag is necessarily set to false owing to the fact that if the
user is using long lived creds then the sessionToken is not required.

Although testing shows `get-session-token` it is anticipated
that the majority of users would use `assume-role` to generate
these temporary credentials.

Fixes #103

Signed-off-by: Richard Gee <richard@technologee.co.uk>