Use explicit path for SUCCESS_CMD
This propagates a change that was initially made in arkade
(alexellis/arkade#685) after it was found
that sudo on CentOS9 was not able to find the new binary in the path
it uses by default. This led to a poor UX as the binary had been
installed but an error in running SUCCESS_CMD gave the impression
that the installation had failed. This change adds an explicit
path to SUCCESS_CMD so that it can be found when running through sudo
with a 'secure_path' set.

Also taken the opportunity to simplify the area that extracts the binary
from the downloaded archive.

Signed-off-by: Richard Gee <richard@technologee.co.uk>