feat: added name and digital ocean flags
The current OS image used for Digital Ocean is no longer available. If
you try to provision a inlets host it fails. This change updates the
default image used and exposes flags so you manually change the OS and
plan if needed.

Also, added a new optional 'name' flag that can be used to explicitly
set the name of the host created, instead of relying on a random name.
This is useful for users that need a name that is more obvious and
attributable to a person.

Signed-off-by: Richard Case <richard.case@outlook.com>